Web App Interface (#1)
What good is a command line app? Let's make a web app and share it with others.

To make the app work within a web context, implements **multi-threaded browsing** to speed things up / avoid timeouts, and enables **in-memory storage** of courses (instead of csv file storage) to enable the app to be run on a server.